backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
django-cors-headers>=4.4,<4.5
python-dotenv>=1.0,<1.1
